fix: repair view_sim unpack and add basedpyright checks.
drop_scores now returns five values; reuse that result instead of simulating twice. Add basic type checking so arity mismatches fail before runtime.

+++ scripts/train_sim.py
@@ -180,7 +180,7 @@ def run_rl(policy: LinearPolicy, cfg: RlConfig) -> None:
         flush=True,
     )
     window: list[float] = []
-    window_steps: list[int] = []
+    window_steps: list[float] = []
     rl_lr = cfg.lr
     entropy = cfg.entropy_coef
     batch_num = 0
